search for continent done

+++ src/main/java/org/posila/cities/cities/dao/CityDAO.java
@@ -1,18 +1,21 @@
 package org.posila.cities.cities.dao;
 
 import org.posila.cities.cities.entities.City;
+import org.posila.cities.cities.entities.Continent;
 import org.posila.cities.cities.entities.Country;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
 import java.util.Collection;
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Component
 public class CityDAO {
     private CityRepository cityRepository;
     private CountryDAO countryDAO;
+    private ContinentDAO continentDAO;
 
     public void saveAll(Iterable<City> cities) {
         cityRepository.saveAll(cities);
@@ -44,13 +47,23 @@ public class CityDAO {
         cityRepository.deleteAll(cities);
     }
 
-    public void deleteAllFromCountry(String countryName, String[] cityNames) {
-        Country country = countryDAO.findByNameOrThrowException(countryName);
+    public String deleteAllFromCountry(String continentName, String countryName, String[] cityNames) {
+        Optional<Continent> optionalContinent = continentDAO.findByName(continentName);
+        if (!optionalContinent.isPresent()) {
+            return String.format("Couldn't find continent %s.", continentName);
+        }
+        Optional<Country> optionalCountry = optionalContinent.get().getCountries()
+                .stream().filter(c -> c.getName().equals(countryName)).findAny();
+        if(!optionalCountry.isPresent()) {
+            return String.format("Couldn't find country %s in continent %s.", countryName, continentName);
+        }
+        Country country = optionalCountry.get();
         cityRepository.deleteAll(
                 cityRepository.findAllByNameIn(cityNames).stream()
                         .filter(city -> country.getCities().remove(city))
                         .collect(Collectors.toList()));
         countryDAO.save(country);
+        return null;
     }
 
     @Autowired
@@ -61,5 +74,10 @@ public class CityDAO {
     @Autowired
     public void setCountryDAO(CountryDAO countryDAO) {
         this.countryDAO = countryDAO;
+    }
+
+    @Autowired
+    public void setContinentDAO(ContinentDAO continentDAO) {
+        this.continentDAO = continentDAO;
     }
 }
